Extract code for creating similar artists to a separate class
